+++ lib/legion/cli/chat/tool_registry.rb
@@ -42,6 +42,7 @@ begin
   require 'legion/cli/chat/tools/entity_extract'
   require 'legion/cli/chat/tools/arbitrage_status'
   require 'legion/cli/chat/tools/escalation_status'
+  require 'legion/cli/chat/tools/graph_explore'
 rescue LoadError => e
   Legion::Logging.debug("ToolRegistry ruby_llm not available, chat tools will not be registered: #{e.message}") if defined?(Legion::Logging)
 end
@@ -90,7 +91,8 @@ module Legion
                             Tools::ShadowEvalStatus,
                             Tools::EntityExtract,
                             Tools::ArbitrageStatus,
-                            Tools::EscalationStatus
+                            Tools::EscalationStatus,
+                            Tools::GraphExplore
                           ].freeze
                         else
                           [].freeze